+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Borehole equivalent resistances is now calculated in one step, centralised in the pipe class (issue #45).
 - Go to 100% code coverage with 300 tests.
 - Threshold interpolation for g-functions set to a relative threshold of 25% relative to the demanded depth (issue #144).
+- _size_based_on_temperature_profile now returns two arguments: the required depth and a boolean flag to check if the field is sized (issue #144).
+- Speed up of L3/L4 sizing by halving calculation time due to intermediate checks if the field is sized (issue #144).
 
 ## Fixed
 - Fixed problem with L2 sizing, when the peak load was the same in all months (issue #146).